refactor(tests): Update E2E AirScript tests to remove WithoutMIR pipeline
