+++ air-script/src/cli/transpile.rs
@@ -19,10 +19,6 @@ impl Target {
     }
 }
 
-
-
-
-
 #[derive(Args)]
 pub struct Transpile {
     /// Path to input file
