+++ air-script/tests/codegen/mod.rs
@@ -1,3 +1,2 @@
 mod helpers;
-mod winterfell_with_mir;
-mod winterfell_wo_mir;
+mod winterfell;
