+++ air-script/tests/codegen/winterfell.rs
@@ -1,6 +1,6 @@
 use expect_test::expect_file;
 
-use super::helpers::{Pipeline, Target, Test};
+use super::helpers::{Target, Test};
 
 // tests_wo_mir
 // ================================================================================================
@@ -8,7 +8,7 @@ use super::helpers::{Pipeline, Target, Test};
 #[test]
 fn binary() {
     let generated_air = Test::new("tests/binary/binary.air".to_string())
-        .transpile(Target::Winterfell, Pipeline::WithoutMIR)
+        .transpile(Target::Winterfell)
         .unwrap();
 
     let expected = expect_file!["../binary/binary.rs"];
@@ -17,22 +17,67 @@ fn binary() {
 
 #[test]
 fn buses_simple() {
-    Test::new("tests/buses/buses_simple.air".to_string())
-        .transpile(Target::Winterfell, Pipeline::WithoutMIR)
-        .expect_err("Buses should not be supported in the WithoutMIR pipeline");
+    let generated_air = Test::new("tests/buses/buses_simple.air".to_string())
+        .transpile(Target::Winterfell)
+        .unwrap();
+
+    let expected = expect_file!["../buses/buses_simple.rs"];
+    expected.assert_eq(&generated_air);
+}
+#[test]
+fn buses_simple_with_evaluators() {
+    let generated_air = Test::new("tests/buses/buses_simple_with_evaluators.air".to_string())
+        .transpile(Target::Winterfell)
+        .unwrap();
+
+    let expected = expect_file!["../buses/buses_simple.rs"];
+    expected.assert_eq(&generated_air);
 }
 
 #[test]
 fn buses_complex() {
-    Test::new("tests/buses/buses_complex.air".to_string())
-        .transpile(Target::Winterfell, Pipeline::WithoutMIR)
-        .expect_err("Buses should not be supported in the WithoutMIR pipeline");
+    let generated_air = Test::new("tests/buses/buses_complex.air".to_string())
+        .transpile(Target::Winterfell)
+        .unwrap();
+
+    let expected = expect_file!["../buses/buses_complex.rs"];
+    expected.assert_eq(&generated_air);
+}
+
+#[test]
+fn buses_varlen_boundary_first() {
+    let generated_air = Test::new("tests/buses/buses_varlen_boundary_first.air".to_string())
+        .transpile(Target::Winterfell)
+        .unwrap();
+
+    let expected = expect_file!["../buses/buses_varlen_boundary_first.rs"];
+    expected.assert_eq(&generated_air);
+}
+
+#[test]
+fn buses_varlen_boundary_last() {
+    let generated_air = Test::new("tests/buses/buses_varlen_boundary_last.air".to_string())
+        .transpile(Target::Winterfell)
+        .unwrap();
+
+    let expected = expect_file!["../buses/buses_varlen_boundary_last.rs"];
+    expected.assert_eq(&generated_air);
+}
+
+#[test]
+fn buses_varlen_boundary_both() {
+    let generated_air = Test::new("tests/buses/buses_varlen_boundary_both.air".to_string())
+        .transpile(Target::Winterfell)
+        .unwrap();
+
+    let expected = expect_file!["../buses/buses_varlen_boundary_both.rs"];
+    expected.assert_eq(&generated_air);
 }
 
 #[test]
 fn periodic_columns() {
     let generated_air = Test::new("tests/periodic_columns/periodic_columns.air".to_string())
-        .transpile(Target::Winterfell, Pipeline::WithoutMIR)
+        .transpile(Target::Winterfell)
         .unwrap();
 
     let expected = expect_file!["../periodic_columns/periodic_columns.rs"];
@@ -42,7 +87,7 @@ fn periodic_columns() {
 #[test]
 fn pub_inputs() {
     let generated_air = Test::new("tests/pub_inputs/pub_inputs.air".to_string())
-        .transpile(Target::Winterfell, Pipeline::WithoutMIR)
+        .transpile(Target::Winterfell)
         .unwrap();
 
     let expected = expect_file!["../pub_inputs/pub_inputs.rs"];
@@ -52,7 +97,7 @@ fn pub_inputs() {
 #[test]
 fn system() {
     let generated_air = Test::new("tests/system/system.air".to_string())
-        .transpile(Target::Winterfell, Pipeline::WithoutMIR)
+        .transpile(Target::Winterfell)
         .unwrap();
 
     let expected = expect_file!["../system/system.rs"];
@@ -62,7 +107,7 @@ fn system() {
 #[test]
 fn bitwise() {
     let generated_air = Test::new("tests/bitwise/bitwise.air".to_string())
-        .transpile(Target::Winterfell, Pipeline::WithoutMIR)
+        .transpile(Target::Winterfell)
         .unwrap();
 
     let expected = expect_file!["../bitwise/bitwise.rs"];
@@ -72,7 +117,7 @@ fn bitwise() {
 #[test]
 fn constants() {
     let generated_air = Test::new("tests/constants/constants.air".to_string())
-        .transpile(Target::Winterfell, Pipeline::WithoutMIR)
+        .transpile(Target::Winterfell)
         .unwrap();
 
     let expected = expect_file!["../constants/constants.rs"];
@@ -82,7 +127,7 @@ fn constants() {
 #[test]
 fn constant_in_range() {
     let generated_air = Test::new("tests/constant_in_range/constant_in_range.air".to_string())
-        .transpile(Target::Winterfell, Pipeline::WithoutMIR)
+        .transpile(Target::Winterfell)
         .unwrap();
 
     let expected = expect_file!["../constant_in_range/constant_in_range.rs"];
@@ -92,7 +137,7 @@ fn constant_in_range() {
 #[test]
 fn evaluators() {
     let generated_air = Test::new("tests/evaluators/evaluators.air".to_string())
-        .transpile(Target::Winterfell, Pipeline::WithoutMIR)
+        .transpile(Target::Winterfell)
         .unwrap();
 
     let expected = expect_file!["../evaluators/evaluators.rs"];
@@ -102,7 +147,7 @@ fn evaluators() {
 #[test]
 fn fibonacci() {
     let generated_air = Test::new("tests/fibonacci/fibonacci.air".to_string())
-        .transpile(Target::Winterfell, Pipeline::WithoutMIR)
+        .transpile(Target::Winterfell)
         .unwrap();
 
     let expected = expect_file!["../fibonacci/fibonacci.rs"];
@@ -112,7 +157,7 @@ fn fibonacci() {
 #[test]
 fn functions_simple() {
     let generated_air = Test::new("tests/functions/functions_simple.air".to_string())
-        .transpile(Target::Winterfell, Pipeline::WithoutMIR)
+        .transpile(Target::Winterfell)
         .unwrap();
 
     let expected = expect_file!["../functions/functions_simple.rs"];
@@ -124,7 +169,7 @@ fn functions_simple_inlined() {
     // make sure that the constraints generated using inlined functions are the same as the ones
     // generated using regular functions
     let generated_air = Test::new("tests/functions/inlined_functions_simple.air".to_string())
-        .transpile(Target::Winterfell, Pipeline::WithoutMIR)
+        .transpile(Target::Winterfell)
         .unwrap();
 
     let expected = expect_file!["../functions/functions_simple.rs"];
@@ -134,7 +179,7 @@ fn functions_simple_inlined() {
 #[test]
 fn functions_complex() {
     let generated_air = Test::new("tests/functions/functions_complex.air".to_string())
-        .transpile(Target::Winterfell, Pipeline::WithoutMIR)
+        .transpile(Target::Winterfell)
         .unwrap();
 
     let expected = expect_file!["../functions/functions_complex.rs"];
@@ -144,7 +189,7 @@ fn functions_complex() {
 #[test]
 fn variables() {
     let generated_air = Test::new("tests/variables/variables.air".to_string())
-        .transpile(Target::Winterfell, Pipeline::WithoutMIR)
+        .transpile(Target::Winterfell)
         .unwrap();
 
     let expected = expect_file!["../variables/variables.rs"];
@@ -154,7 +199,7 @@ fn variables() {
 #[test]
 fn trace_col_groups() {
     let generated_air = Test::new("tests/trace_col_groups/trace_col_groups.air".to_string())
-        .transpile(Target::Winterfell, Pipeline::WithoutMIR)
+        .transpile(Target::Winterfell)
         .unwrap();
 
     let expected = expect_file!["../trace_col_groups/trace_col_groups.rs"];
@@ -165,7 +210,7 @@ fn trace_col_groups() {
 fn indexed_trace_access() {
     let generated_air =
         Test::new("tests/indexed_trace_access/indexed_trace_access.air".to_string())
-            .transpile(Target::Winterfell, Pipeline::WithoutMIR)
+            .transpile(Target::Winterfell)
             .unwrap();
 
     let expected = expect_file!["../indexed_trace_access/indexed_trace_access.rs"];
@@ -175,7 +220,7 @@ fn indexed_trace_access() {
 #[test]
 fn list_comprehension() {
     let generated_air = Test::new("tests/list_comprehension/list_comprehension.air".to_string())
-        .transpile(Target::Winterfell, Pipeline::WithoutMIR)
+        .transpile(Target::Winterfell)
         .unwrap();
 
     let expected = expect_file!["../list_comprehension/list_comprehension.rs"];
@@ -183,9 +228,20 @@ fn list_comprehension() {
 }
 
 #[test]
+fn list_comprehension_nested() {
+    let generated_air =
+        Test::new("tests/list_comprehension/list_comprehension_nested.air".to_string())
+            .transpile(Target::Winterfell)
+            .unwrap();
+
+    let expected = expect_file!["../list_comprehension/list_comprehension_nested.rs"];
+    expected.assert_eq(&generated_air);
+}
+
+#[test]
 fn list_folding() {
     let generated_air = Test::new("tests/list_folding/list_folding.air".to_string())
-        .transpile(Target::Winterfell, Pipeline::WithoutMIR)
+        .transpile(Target::Winterfell)
         .unwrap();
 
     let expected = expect_file!["../list_folding/list_folding.rs"];
@@ -195,17 +251,17 @@ fn list_folding() {
 #[test]
 fn selectors() {
     let generated_air = Test::new("tests/selectors/selectors.air".to_string())
-        .transpile(Target::Winterfell, Pipeline::WithoutMIR)
+        .transpile(Target::Winterfell)
         .unwrap();
 
     let expected = expect_file!["../selectors/selectors.rs"];
     expected.assert_eq(&generated_air);
 
     let generated_air = Test::new("tests/selectors/selectors_with_evaluators.air".to_string())
-        .transpile(Target::Winterfell, Pipeline::WithoutMIR)
+        .transpile(Target::Winterfell)
         .unwrap();
 
-    let expected = expect_file!["../selectors/selectors.rs"];
+    let expected = expect_file!["../selectors/selectors_with_evaluators.rs"];
     expected.assert_eq(&generated_air);
 }
 
@@ -213,7 +269,7 @@ fn selectors() {
 fn constraint_comprehension() {
     let generated_air =
         Test::new("tests/constraint_comprehension/constraint_comprehension.air".to_string())
-            .transpile(Target::Winterfell, Pipeline::WithoutMIR)
+            .transpile(Target::Winterfell)
             .unwrap();
 
     let expected = expect_file!["../constraint_comprehension/constraint_comprehension.rs"];
@@ -221,7 +277,7 @@ fn constraint_comprehension() {
 
     let generated_air =
         Test::new("tests/constraint_comprehension/cc_with_evaluators.air".to_string())
-            .transpile(Target::Winterfell, Pipeline::WithoutMIR)
+            .transpile(Target::Winterfell)
             .unwrap();
 
     let expected = expect_file!["../constraint_comprehension/constraint_comprehension.rs"];
