+++ air-script/tests/functions/mod.rs
@@ -3,8 +3,5 @@
 mod functions_complex;
 #[rustfmt::skip]
 #[allow(clippy::all)]
-mod functions_complex_with_mir;
-#[rustfmt::skip]
-#[allow(clippy::all)]
 mod functions_simple;
 mod test_air;
