+++ air-script/tests/list_comprehension/mod.rs
@@ -1,7 +1,4 @@
 #[rustfmt::skip]
 #[allow(clippy::all)]
 mod list_comprehension;
-#[rustfmt::skip]
-#[allow(clippy::all)]
-mod list_comprehension_with_mir;
 mod test_air;
