+++ air-script/tests/list_folding/list_folding.rs
@@ -82,10 +82,10 @@ impl Air for ListFoldingAir {
     fn evaluate_transition<E: FieldElement<BaseField = Felt>>(&self, frame: &EvaluationFrame<E>, periodic_values: &[E], result: &mut [E]) {
         let main_current = frame.current();
         let main_next = frame.next();
-        result[0] = main_next[5] - (main_current[9] + main_current[10] + main_current[11] + main_current[12] + main_current[13] * main_current[14] * main_current[15] * main_current[16]);
-        result[1] = main_next[6] - (main_current[9] + main_current[10] + main_current[11] + main_current[12] + main_current[13] * main_current[14] * main_current[15] * main_current[16]);
-        result[2] = main_next[7] - (main_current[9] * main_current[13] + main_current[10] * main_current[14] + main_current[11] * main_current[15] + main_current[12] * main_current[16] + (main_current[9] + main_current[13]) * (main_current[10] + main_current[14]) * (main_current[11] + main_current[15]) * (main_current[12] + main_current[16]));
-        result[3] = main_next[8] - (main_current[1] + main_current[9] * main_current[13] + main_current[10] * main_current[14] + main_current[11] * main_current[15] + main_current[12] * main_current[16] + main_current[9] * main_current[13] + main_current[10] * main_current[14] + main_current[11] * main_current[15] + main_current[12] * main_current[16]);
+        result[0] = main_next[5] - (E::ZERO + main_current[9] + main_current[10] + main_current[11] + main_current[12] + E::ONE * main_current[13] * main_current[14] * main_current[15] * main_current[16]);
+        result[1] = main_next[6] - (E::ZERO + main_current[9] + main_current[10] + main_current[11] + main_current[12] + E::ONE * main_current[13] * main_current[14] * main_current[15] * main_current[16]);
+        result[2] = main_next[7] - (E::ZERO + main_current[9] * main_current[13] + main_current[10] * main_current[14] + main_current[11] * main_current[15] + main_current[12] * main_current[16] + E::ONE * (main_current[9] + main_current[13]) * (main_current[10] + main_current[14]) * (main_current[11] + main_current[15]) * (main_current[12] + main_current[16]));
+        result[3] = main_next[8] - (main_current[1] + E::ZERO + main_current[9] * main_current[13] + main_current[10] * main_current[14] + main_current[11] * main_current[15] + main_current[12] * main_current[16] + E::ZERO + main_current[9] * main_current[13] + main_current[10] * main_current[14] + main_current[11] * main_current[15] + main_current[12] * main_current[16]);
     }
 
     fn evaluate_aux_transition<F, E>(&self, main_frame: &EvaluationFrame<F>, aux_frame: &EvaluationFrame<E>, _periodic_values: &[F], aux_rand_elements: &AuxRandElements<E>, result: &mut [E])
