+++ air-script/tests/list_folding/mod.rs
@@ -1,7 +1,4 @@
 #[rustfmt::skip]
 #[allow(clippy::all)]
 mod list_folding;
-#[rustfmt::skip]
-#[allow(clippy::all)]
-mod list_folding_with_mir;
 mod test_air;
