+++ air-script/tests/selectors/mod.rs
@@ -4,10 +4,4 @@ mod selectors;
 #[rustfmt::skip]
 #[allow(clippy::all)]
 mod selectors_with_evaluators;
-#[rustfmt::skip]
-#[allow(clippy::all)]
-mod selectors_with_evaluators_with_mir;
-#[rustfmt::skip]
-#[allow(clippy::all)]
-mod selectors_with_mir;
 mod test_air;
